// frontend/src/lib/api/generated/model/mediaFolder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Construction ERP API
 * OpenAPI spec version: 0.1.0
 */

export type MediaFolder = typeof MediaFolder[keyof typeof MediaFolder];


export const MediaFolder = {
  drawings: 'drawings',
  contracts: 'contracts',
  permits: 'permits',
  photos: 'photos',
  reports: 'reports',
  other: 'other',
} as const;